refactor: query/command package 분리
