+++ src/main/java/com/groom/orbit/goal/controller/command/QuestCommandController.java
@@ -1,4 +1,4 @@
-package com.groom.orbit.goal.controller;
+package com.groom.orbit.goal.controller.command;
 
 import java.util.List;
 
@@ -13,8 +13,8 @@ import org.springframework.web.bind.annotation.RestController;
 import com.groom.orbit.common.annotation.AuthMember;
 import com.groom.orbit.common.dto.CommonSuccessDto;
 import com.groom.orbit.common.dto.ResponseDto;
-import com.groom.orbit.goal.app.QuestCommandService;
-import com.groom.orbit.goal.app.QuestUpdateService;
+import com.groom.orbit.goal.app.command.QuestCommandService;
+import com.groom.orbit.goal.app.command.QuestUpdateService;
 import com.groom.orbit.goal.app.dto.CreateQuestRequestDto;
 import com.groom.orbit.goal.app.dto.UpdateQuestRequestDto;
 import com.groom.orbit.goal.app.dto.UpdateQuestSequenceRequestDto;
