+++ src/main/java/com/groom/orbit/goal/controller/query/GoalQueryController.java
@@ -1,12 +1,12 @@
-package com.groom.orbit.goal.controller;
+package com.groom.orbit.goal.controller.query;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.groom.orbit.common.dto.ResponseDto;
-import com.groom.orbit.goal.app.GoalQueryService;
 import com.groom.orbit.goal.app.dto.GetGoalCategoryResponseDto;
+import com.groom.orbit.goal.app.query.GoalQueryService;
 
 import lombok.RequiredArgsConstructor;
